files modified

+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: mamaquig <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/09 18:39:13 by mamaquig          #+#    #+#             */
-/*   Updated: 2019/10/11 18:42:55 by mamaquig         ###   ########.fr       */
+/*   Updated: 2019/10/21 18:39:22 by mamaquig         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int i;
-	int j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	if (needle[0] == 0)
